Top-up: complete the screens, and fix three /payment/interest defects
Built out the sections the port was missing, against the QA reference
(LandAndHouseWeb/docs/topup-user-manual-v2.md and its screenshots):

- card: the วิธีขอปรับวงเงินเพิ่ม conditions panel, the
  รับเงินโอนเข้าบัญชีสูงสุด band, and the full credit summary
  (credit_limit / current_ltv_amount / default_topup_amount /
  balance_receivable at data_date / default_transfer_amount)
- amount: the numbered deduction list 1, 2, 3, (4), 5 with 5.1/5.2 and
  per-row contract numbers, plus ปรับปรุงยอดชำระ
- photos: retitled ข้อมูลการต่อภาษี, with the registration/tax block
- conclusion: the สรุปยอดสินเชื่อใหม่ section
- success: the payout/deadline caveat and ดูสถานะการขอเพิ่มวงเงิน
- QR: payment-window note, R1/R2, bank exclusions, ปรับปรุงยอดชำระ

POST /payment/interest was returning 500 where the same call succeeded
from Postman. Three mismatches, none of which produce a 400 because the
request is well-formed:

- comcode came from contract_details, not barcode_details — a different
  field, and the one that identifies the biller
- firstname/lastname came from the app user's profile rather than being
  split off the contract holder's name
- interest_amount, collection_fee and penalty_fee went out as ints

That last one was a porting defect wider than this endpoint: yield,
collection_fee and penalty_fee are decimals on the wire and were parsed
as int on both LoanAmountDetail and LoanContract.topupDetail. 2987.84
became 2987 — a misstated amount billed against an exact figure, and the
wrong JSON type — and it silently corrupted the QR total and the
deduction rows too. Now double throughout. No P-Loan code reads those
three, so no P-Loan behaviour changed.

The name split is also made safe: the source's split(' ')[1] throws
RangeError on a one-word contract name and drops the third token of a
three-part one.

Step 5 now captures with image_picker instead of the host camera bridge.
The host applies its ID-card framing mask to every action but 'selfie',
which is the wrong frame for a vehicle, a tax disc or an odometer, and it
cannot be taught new masks without an app release. Step 7 keeps the
bridge, where the masks are right; a test pins the split both ways.

That dropped the bridge's native downscale, and image_picker_for_web
ignores maxWidth/imageQuality — seven full-resolution photos base64'd
into one JSON body would be tens of megabytes. services/image_downscale
re-encodes each capture through a canvas and toDataURL, the browser's own
JPEG encoder; dart:ui cannot encode JPEG and package:image takes seconds
per 12MP photo in a WebView. It returns the original bytes on any
failure, since an oversized photo beats a lost one.

Left out deliberately: บันทึกรูปภาพ on the QR screen. The source saves
via a native action with no equivalent here, and a web download is not
reliably honoured in the WebView — a button that silently does nothing is
worse than none.

analyze holds at the 39-info baseline; 282 tests pass; web release build
clean. The interest fix is unverified end to end — no token or live
contract available here.

+++ lib/p_loan/application/models/loan_amount_detail.dart
@@ -84,8 +84,12 @@ class LoanAmountDetail {
   /// Stamp duty, deducted from the payout.
   final int feeAmount;
   final int balanceReceivable;
-  final int collectionFee;
-  final int penaltyFee;
+  /// ⚠ **Decimal on the wire**, and it matters: `/payment/interest` is billed
+  /// for an exact outstanding figure, so truncating `38.00` or `2987.84` to an
+  /// int both misstates the amount and changes the JSON type the server is
+  /// given.
+  final double collectionFee;
+  final double penaltyFee;
   final double overdueAmount;
   final String overdueFrom;
   final String overdueTo;
@@ -95,7 +99,9 @@ class LoanAmountDetail {
   final String interestPaidFlag;
 
   /// `yield` on the wire — accrued interest (renamed: Dart reserved word).
-  final int interestYield;
+  ///
+  /// ⚠ **Decimal**, for the same reason as [collectionFee].
+  final double interestYield;
   final int topupSpecials;
 
   bool get isOk => code == '200';
@@ -175,13 +181,13 @@ class LoanAmountDetail {
         topupActual: asInt(json['topup_actual']),
         feeAmount: asInt(json['fee_amount']),
         balanceReceivable: asInt(json['balance_receivable']),
-        collectionFee: asInt(json['collection_fee']),
-        penaltyFee: asInt(json['penalty_fee']),
+        collectionFee: asDouble(json['collection_fee']),
+        penaltyFee: asDouble(json['penalty_fee']),
         overdueAmount: asDouble(json['overdue_amount']),
         overdueFrom: asString(json['overdue_from']),
         overdueTo: asString(json['overdue_to']),
         interestPaidFlag: asString(json['interest_paid_flag']),
-        interestYield: asInt(json['yield']),
+        interestYield: asDouble(json['yield']),
         topupSpecials: asInt(json['topup_specials']),
       );
